Refactors HTTP utility and verification code for better maintainability
Restructures fetch function parameters using destructuring to improve API clarity and adds proper parameter validation.

Extracts verification utilities into a dedicated module to enable code reuse across scripts.

Migrates YAML generation from manual string concatenation to proper YAML library serialization, eliminating potential escaping issues and improving reliability.
