+++ package.json
@@ -36,7 +36,8 @@
     "astro": "^5.13.3",
     "minisearch": "^7.1.2",
     "fflate": "^0.8.1",
-    "sirv-cli": "^3.0.0"
+    "sirv-cli": "^3.0.0",
+    "yaml": "^2.5.1"
   },
   "devDependencies": {
     "@astrojs/check": "^0.9.4",
